Milestone 1.2: shared-memory channels, with tests
Code:
- channels.py: heartbeats as c_double, since time.monotonic() is a float
- cli.py: spawn context; channels created as lock-free RawValue shared
  memory inside the lock, then handed to the app

Tests:
- test_channels.py (the 1.2 gate): a spawned child's increments reach
  the parent, a kill -9 leaves totals at their last value, a child's
  heartbeat arrives fresh, and the stop flag reaches the child
- test_cli.py: Pulse.run replaced by a recording stand-in, so the
  startup path runs without the TUI; the app runs once under the lock
  and receives channels in shared memory
- conftest.py: lock_is_free and is_shared shared across test modules

+++ src/pulse/channels.py
@@ -1,5 +1,5 @@
 import ctypes
-from datalasses import dataclass
+from dataclasses import dataclass
 
 
 class Counter(ctypes.Structure):
@@ -11,8 +11,8 @@ class Counter(ctypes.Structure):
 
 class Heartbeat(ctypes.Structure):
     _fields_ = [
-        ("ingest", ctypes.c_int),
-        ("etl", ctypes.c_int),
+        ("ingest", ctypes.c_double),
+        ("etl", ctypes.c_double),
     ]
